+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/extension/WebExtensionPromptFeature.kt
@@ -6,6 +6,10 @@ package org.mozilla.fenix.extension
 
 import android.content.Context
 import android.view.Gravity
+import android.view.LayoutInflater
+import android.view.View
+import android.widget.TextView
+import androidx.annotation.UiContext
 import androidx.annotation.VisibleForTesting
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.FragmentManager
@@ -27,6 +31,7 @@ import mozilla.components.lib.state.ext.flowScoped
 import mozilla.components.support.base.feature.LifecycleAwareFeature
 import mozilla.components.support.ktx.android.content.appVersionName
 import mozilla.components.ui.widgets.withCenterAlignedButtons
+import org.mozilla.fenix.BuildConfig
 import org.mozilla.fenix.R
 import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.theme.ThemeManager
@@ -36,9 +41,10 @@ import org.mozilla.fenix.theme.ThemeManager
  */
 class WebExtensionPromptFeature(
     private val store: BrowserStore,
-    private val context: Context,
-    private val addonManager: AddonManager = context.components.addonManager,
+    @UiContext private val context: Context,
     private val fragmentManager: FragmentManager,
+    private val onLinkClicked: (String, Boolean) -> Unit,
+    private val addonManager: AddonManager = context.components.addonManager,
 ) : LifecycleAwareFeature {
 
     /**
@@ -148,21 +154,24 @@ class WebExtensionPromptFeature(
     @VisibleForTesting
     internal fun handleInstallationFailedRequest(
         exception: WebExtensionInstallException,
-    ) {
+    ): AlertDialog? {
         val addonName = exception.extensionName ?: ""
         var title = context.getString(R.string.mozac_feature_addons_cant_install_extension)
+        var url: String? = null
         val message = when (exception) {
             is WebExtensionInstallException.Blocklisted -> {
+                url = exception.extensionId?.let { AMO_BLOCKED_PAGE_URL.format(it) }
                 context.getString(R.string.mozac_feature_addons_blocklisted_1, addonName)
             }
 
             is WebExtensionInstallException.SoftBlocked -> {
+                url = exception.extensionId?.let { AMO_BLOCKED_PAGE_URL.format(it) }
                 context.getString(R.string.mozac_feature_addons_soft_blocked, addonName)
             }
 
             is WebExtensionInstallException.UserCancelled -> {
                 // We don't want to show an error message when users cancel installation.
-                return
+                return null
             }
 
             is WebExtensionInstallException.UnsupportedAddonType,
@@ -209,9 +218,10 @@ class WebExtensionPromptFeature(
             }
         }
 
-        showDialog(
+        return showDialog(
             title = title,
             message = message,
+            url = url,
         )
     }
 
@@ -379,17 +389,34 @@ class WebExtensionPromptFeature(
     internal fun showDialog(
         title: String,
         message: String,
-    ) {
+        url: String? = null,
+    ): AlertDialog {
         context.let {
-            AlertDialog.Builder(it)
+            var dialog: AlertDialog? = null
+            val inflater = LayoutInflater.from(it)
+            val view = inflater.inflate(R.layout.addon_installation_failed_dialog, null)
+            val messageView = view.findViewById<TextView>(R.id.message)
+            messageView.text = message
+
+            if (url != null) {
+                val linkView = view.findViewById<TextView>(R.id.link)
+                linkView.visibility = View.VISIBLE
+                linkView.setOnClickListener {
+                    onLinkClicked(url, true) // shouldOpenInBrowser
+                    dialog?.dismiss()
+                }
+            }
+
+            dialog = AlertDialog.Builder(it)
                 .setTitle(title)
                 .setPositiveButton(android.R.string.ok) { _, _ -> }
                 .setCancelable(false)
-                .setMessage(
-                    message,
-                )
-                .show()
+                .setView(view)
+                .create()
                 .withCenterAlignedButtons()
+            dialog.show()
+
+            return dialog
         }
     }
 
@@ -397,5 +424,6 @@ class WebExtensionPromptFeature(
         private const val PERMISSIONS_DIALOG_FRAGMENT_TAG = "ADDONS_PERMISSIONS_DIALOG_FRAGMENT"
         private const val POST_INSTALLATION_DIALOG_FRAGMENT_TAG =
             "ADDONS_INSTALLATION_DIALOG_FRAGMENT"
+        private const val AMO_BLOCKED_PAGE_URL = "${BuildConfig.AMO_BASE_URL}/android/blocked-addon/%s/"
     }
 }