+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/extension/WebExtensionPromptFeatureTest.kt
@@ -4,6 +4,7 @@
 
 package org.mozilla.fenix.extension
 
+import androidx.core.view.isVisible
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
@@ -19,18 +20,24 @@ import mozilla.components.support.ktx.android.content.appVersionName
 import mozilla.components.support.test.ext.joinBlocking
 import mozilla.components.support.test.robolectric.testContext
 import mozilla.components.support.test.rule.MainCoroutineRule
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mozilla.fenix.BuildConfig
 import org.mozilla.fenix.R
 import org.mozilla.fenix.helpers.FenixRobolectricTestRunner
+import org.mozilla.fenix.utils.LinkTextView
 
 @RunWith(FenixRobolectricTestRunner::class)
 class WebExtensionPromptFeatureTest {
 
     private lateinit var webExtensionPromptFeature: WebExtensionPromptFeature
     private lateinit var store: BrowserStore
+
+    private val onLinkClicked: (String, Boolean) -> Unit = spyk()
 
     @get:Rule
     val coroutinesTestRule = MainCoroutineRule()
@@ -43,6 +50,7 @@ class WebExtensionPromptFeatureTest {
                 store = store,
                 context = testContext,
                 fragmentManager = mockk(relaxed = true),
+                onLinkClicked = onLinkClicked,
                 addonManager = mockk(relaxed = true),
             ),
         )
@@ -52,7 +60,7 @@ class WebExtensionPromptFeatureTest {
     fun `WHEN InstallationFailed is dispatched THEN handleInstallationFailedRequest is called`() {
         webExtensionPromptFeature.start()
 
-        every { webExtensionPromptFeature.handleInstallationFailedRequest(any()) } just runs
+        every { webExtensionPromptFeature.handleInstallationFailedRequest(any()) } returns null
 
         store.dispatch(
             UpdatePromptRequestWebExtensionAction(
@@ -79,29 +87,39 @@ class WebExtensionPromptFeatureTest {
                 "name",
             )
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
     fun `WHEN calling handleInstallationFailedRequest with Blocklisted error THEN showDialog with the correct message`() {
         val expectedTitle = testContext.getString(R.string.mozac_feature_addons_cant_install_extension)
+        val extensionId = "extensionId"
         val extensionName = "extensionName"
         val exception = WebExtensionInstallException.Blocklisted(
+            extensionId = extensionId,
             extensionName = extensionName,
             throwable = Exception(),
         )
         val expectedMessage =
             testContext.getString(R.string.mozac_feature_addons_blocklisted_1, extensionName)
+        val expectedUrl = "${BuildConfig.AMO_BASE_URL}/android/blocked-addon/$extensionId/"
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
-        verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage, expectedUrl) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertTrue(linkView!!.isVisible)
+
+        // Click the link, then verify.
+        linkView.performClick()
+        verify {
+            onLinkClicked(expectedUrl, true)
+            dialog.dismiss()
+        }
     }
 
     @Test
@@ -133,11 +151,11 @@ class WebExtensionPromptFeatureTest {
         val expectedMessage =
             testContext.getString(R.string.mozac_feature_addons_failed_to_install, extensionName)
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
@@ -150,11 +168,11 @@ class WebExtensionPromptFeatureTest {
         val expectedMessage =
             testContext.getString(R.string.mozac_feature_addons_extension_failed_to_install)
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
@@ -166,11 +184,11 @@ class WebExtensionPromptFeatureTest {
         val expectedMessage =
             testContext.getString(R.string.mozac_feature_addons_extension_failed_to_install_corrupt_error)
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
@@ -182,11 +200,11 @@ class WebExtensionPromptFeatureTest {
         val expectedMessage =
             testContext.getString(R.string.mozac_feature_addons_extension_failed_to_install_not_signed_error)
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
@@ -207,11 +225,11 @@ class WebExtensionPromptFeatureTest {
                 version,
             )
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
@@ -314,11 +332,11 @@ class WebExtensionPromptFeatureTest {
         val expectedMessage =
             testContext.getString(R.string.mozac_feature_addons_failed_to_install, extensionName)
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
@@ -332,25 +350,37 @@ class WebExtensionPromptFeatureTest {
         val expectedMessage =
             testContext.getString(R.string.mozac_feature_addons_admin_install_only, extensionName)
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(
-            exception = exception,
-        )
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
         verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertFalse(linkView!!.isVisible)
     }
 
     @Test
     fun `WHEN calling handleInstallationFailedRequest with SoftBlocked error THEN showDialog with the correct message`() {
         val expectedTitle = testContext.getString(R.string.mozac_feature_addons_cant_install_extension)
+        val extensionId = "extensionId"
         val extensionName = "extensionName"
         val exception = WebExtensionInstallException.SoftBlocked(
+            extensionId = extensionId,
             extensionName = extensionName,
             throwable = Exception(),
         )
         val expectedMessage = testContext.getString(R.string.mozac_feature_addons_soft_blocked, extensionName)
+        val expectedUrl = "${BuildConfig.AMO_BASE_URL}/android/blocked-addon/$extensionId/"
 
-        webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
+        val dialog = webExtensionPromptFeature.handleInstallationFailedRequest(exception = exception)
 
-        verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage) }
+        verify { webExtensionPromptFeature.showDialog(expectedTitle, expectedMessage, expectedUrl) }
+        val linkView = dialog?.findViewById<LinkTextView>(R.id.link)
+        assertTrue(linkView!!.isVisible)
+
+        // Click the link, then verify.
+        linkView.performClick()
+        verify {
+            onLinkClicked(expectedUrl, true)
+            dialog.dismiss()
+        }
     }
 }